Move monitor to pkg/util/sync
