+++ pkg/util/sync/monitor.go
@@ -1,4 +1,4 @@
-package util
+package sync
 
 import "sync"
